requirements.txt
# Core Django
Django==5.1.3
django-ninja==1.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt>=5.3.0
drf-spectacular==0.27.2  # OpenAPI/Swagger documentation

# Database
PyMySQL==1.1.1
dj-database-url==2.3.0
cryptography==43.0.3  # Often needed for modern DB connections

# Background Tasks
redis==5.2.0
celery==5.4.0  # Replaces huey for better performance

# Utilities
python-dotenv==1.0.1
pydantic>=2.0.0
Pillow==11.0.0
requests==2.32.3
stripe>=7.0.0
reportlab>=4.0.0  # PDF generation
weasyprint>=60.0  # HTML to PDF generation

# CORS
django-cors-headers==4.6.0

# Static files
whitenoise==6.8.2  # Replaces dj-static/static3 for modern Django

# WSGI/ASGI Server
gunicorn==23.0.0
uvicorn==0.32.1  # Added for ASGI support (Django Ninja/Async)
